Crate rtcp

Package rtcp implements encoding and decoding of RTCP packets according to RFCs 3550 and 5506.

RTCP is a sister protocol of the Real-time Transport Protocol (RTP). Its basic functionality and packet structure is defined in RFC 3550. RTCP provides out-of-band statistics and control information for an RTP session. It partners with RTP in the delivery and packaging of multimedia data, but does not transport any media data itself.

The primary function of RTCP is to provide feedback on the quality of service (QoS) in media distribution by periodically sending statistics information such as transmitted octet and packet counts, packet loss, packet delay variation, and round-trip delay time to participants in a streaming multimedia session. An application may use this information to control quality of service parameters, perhaps by limiting flow, or using a different codec.

Decoding RTCP packets:

     let pkt = rtcp::unmarshal(&rtcp_data).unwrap();

     if let Some(e) = pkt
          .as_any()
          .downcast_ref::<PictureLossIndication>()
      {

      }
     else if let Some(e) = packet
          .as_any()
          .downcast_ref::<Goodbye>(){}
     ....

Encoding RTCP packets:

     let pkt = PictureLossIndication{
         sender_ssrc: sender_ssrc,
         media_ssrc: media_ssrc
     };

     let pli_data = pkt.marshal().unwrap();
     // ...

Modules

compound_packet

extended_report

goodbye

header

packet

payload_feedbacks

raw_packet

receiver_report

reception_report

sender_report

source_description

transport_feedbacks

Enums

Error