rtcp

Module extended_report

Re-exports

pub use dlrr::DLRRReport;

pub use dlrr::DLRRReportBlock;

pub use prt::PacketReceiptTimesReportBlock;

pub use rle::Chunk;

pub use rle::ChunkType;

pub use rle::DuplicateRLEReportBlock;

pub use rle::LossRLEReportBlock;

pub use rle::RLEReportBlock;

pub use rrt::ReceiverReferenceTimeReportBlock;

pub use ssr::StatisticsSummaryReportBlock;

pub use ssr::TTLorHopLimitType;

pub use unknown::UnknownReportBlock;

pub use vm::VoIPMetricsReportBlock;

Modules

dlrr

prt

rle

rrt

ssr

unknown

vm

Structs

ExtendedReport

The ExtendedReport packet is an Implementation of RTCP Extended reports defined in RFC 3611. It is used to convey detailed information about an RTP stream. Each packet contains one or more report blocks, each of which conveys a different kind of information.

XRHeader

XRHeader defines the common fields that must appear at the start of each report block. In typical cases, users of ExtendedReports shouldn’t need to access this. For locally-constructed report blocks, these values will not be accurate until the corresponding packet is marshaled.

Enums

BlockType

BlockType specifies the type of report in a report block Extended Report block types from RFC 3611.

Type Aliases

TypeSpecificField

TypeSpecificField as described in RFC 3611 section 4.5. In typical cases, users of ExtendedReports shouldn’t need to access this, and should instead use the corresponding fields in the actual report blocks themselves.