rtcp

Module header

Structs

Header

A Header is the common header shared by all RTCP packets

Enums

PacketType

PacketType specifies the type of an RTCP packet RTCP packet types registered with IANA. See: https://www.iana.org/assignments/rtp-parameters/rtp-parameters.xhtml#rtp-parameters-4

Constants

COUNT_MASK

COUNT_MAX

COUNT_SHIFT

FORMAT_FIR

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

FORMAT_PLI

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

FORMAT_REMB

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

FORMAT_RRR

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

FORMAT_SLI

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

FORMAT_TCC

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here. https://tools.ietf.org/html/draft-holmer-rmcat-transport-wide-cc-extensions-01#page-5

FORMAT_TLN

Transport and Payload specific feedback messages overload the count field to act as a message type. those are listed here

HEADER_LENGTH

PADDING_MASK

PADDING_SHIFT

RTP_VERSION

SDES_MAX_OCTET_COUNT

SSRC_LENGTH

VERSION_MASK

VERSION_SHIFT